ajout navbar et header sur mobile avec lien fonctionnel
