+++ frontend/package.json
@@ -2,11 +2,14 @@
   "scripts": {
     "dev": "vite --host",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "prettierC": "prettier -c",
+    "prettierW": "prettier -w"
   },
   "dependencies": {
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.7",
     "@mui/material": "^5.15.6",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
