+++ frontend/src/App.jsx
@@ -1,9 +1,13 @@
+import Header from "./components/Header/Header";
+import Navbar from "./components/Navbar/Navbar";
 import Home from "./pages/Home";
 
 function App() {
   return (
     <div>
+      <Header />
       <Home />
+      <Navbar />
     </div>
   );
 }
