+++ frontend/src/components/CandleHome.jsx
@@ -1,4 +1,5 @@
 import { useState, useEffect } from "react";
+import { Link } from "react-router-dom";
 import Card from "@mui/material/Card";
 import CardContent from "@mui/material/CardContent";
 import CardMedia from "@mui/material/CardMedia";
@@ -15,36 +16,48 @@ export default function CandleHome() {
   return (
     <section>
       <h2>Mes Bougies</h2>
+
       <Box
         sx={{
           display: "flex",
           flexDirection: "row",
           flexWrap: "wrap",
-          gap: 5,
+          justifyContent: "center",
+          alignItems: "center",
+          gap: 2,
+          margin: 2,
+          marginBottom: 10,
         }}
       >
         {candles.map((candle) => (
-          <Card sx={{ maxWidth: 345, minWidth: 345 }}>
-            <CardMedia
-              sx={{ height: 140 }}
-              image={candle.imageUrl}
-              title="green iguana"
-            />
-            <CardContent>
-              <Typography gutterBottom variant="h5" component="div">
-                {candle.candleName}
-              </Typography>
-              <Typography variant="body2" color="text.secondary">
-                {candle.candleDescription}
-              </Typography>
-              <Typography variant="body2" color="text.secondary">
-                {candle.candleDescription}
-              </Typography>
-              <Typography variant="body2" color="text.secondary">
-                {candle.candleDescription}
-              </Typography>
-            </CardContent>
-          </Card>
+          <Link to={`/candle/${candle.id}`}>
+            <Card
+              sx={{
+                maxWidth: 100,
+                minWidth: 180,
+                maxHeight: 100,
+                minHeight: 200,
+                display: "flex",
+                flexDirection: "column",
+                justifyContent: "center",
+                alignItems: "center",
+              }}
+            >
+              <CardMedia
+                sx={{ height: 140 }}
+                image={candle.imageUrl}
+                title="green iguana"
+              />
+              <CardContent>
+                <Typography gutterBottom variant="h5" component="div">
+                  {candle.candleName}
+                </Typography>
+                <Typography variant="p" color="text.secondary">
+                  {candle.candleDescription}
+                </Typography>
+              </CardContent>
+            </Card>
+          </Link>
         ))}
       </Box>
     </section>
