+++ frontend/src/components/Header/Header.module.css
@@ -1,0 +1,39 @@
+* {
+  margin: 0;
+}
+
+.navBar {
+  font-weight: bold;
+  display: flex;
+  justify-content: space-around;
+  position: fixed;
+  bottom: 0;
+  width: 100%;
+}
+.container {
+  height: 6rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  position: relative;
+  overflow: hidden;
+  margin-bottom: 0.5rem;
+  text-align: center;
+  width: 100%;
+  margin-bottom: 2rem;
+  background-color: #fefee2;
+}
+
+.svgLogo {
+  margin: 2rem;
+  height: 3rem;
+}
+
+.svgProfil {
+  height: 2rem;
+  margin: 2rem;
+}
+
+.svgLink {
+  text-decoration: none;
+}
