+++ frontend/src/pages/AdminPage.jsx
@@ -1,9 +1,13 @@
 import Admin from "../components/Admin/Admin";
+import Header from "../components/Header/Header";
+import Navbar from "../components/Navbar/Navbar";
 
 function AdminPage() {
   return (
     <div>
+      <Header />
       <Admin />
+      <Navbar />
     </div>
   );
 }
