+++ frontend/src/pages/CandlePage.jsx
@@ -1,9 +1,13 @@
 import CandleById from "../components/CandleById";
+import Header from "../components/Header/Header";
+import Navbar from "../components/Navbar/Navbar";
 
 function CandlePage() {
   return (
     <div>
+      <Header />
       <CandleById />
+      <Navbar />
     </div>
   );
 }
