+++ frontend/src/pages/Home.jsx
@@ -4,7 +4,6 @@ function Home() {
   return (
     <div>
       <CandleHome />
-      <h1>Candle</h1>
     </div>
   );
 }
