+++ frontend/src/pages/LogInPage.jsx
@@ -1,144 +1,15 @@
-import * as React from "react";
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import Button from "@mui/material/Button";
-import CssBaseline from "@mui/material/CssBaseline";
-import TextField from "@mui/material/TextField";
-import Paper from "@mui/material/Paper";
-import Box from "@mui/material/Box";
-import Grid from "@mui/material/Grid";
-import Typography from "@mui/material/Typography";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
+import Header from "../components/Header/Header";
+import Login from "../components/LogIn";
+import Navbar from "../components/Navbar/Navbar";
 
-const defaultTheme = createTheme();
-
-export default function LoginPage() {
-  // États pour le mot de passe et la confirmation du mot de passe
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
-
-  // Hook pour la navigation
-  const navigate = useNavigate();
-
-  // Gestionnaire de changement de l'email
-  const handleMailChange = (event) => {
-    setEmail(event.target.value);
-  };
-  // Gestionnaire de changement du mot de passe
-  const handlePasswordChange = (event) => {
-    setPassword(event.target.value);
-  };
-
-  // Gestionnaire de soumission du formulaire
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-
-    try {
-      // Appel à l'API pour créer un nouvel utilisateur
-      const response = await fetch(
-        `${import.meta.env.VITE_BACKEND_URL}/api/login`,
-        {
-          method: "post",
-          credentials: "include",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            email,
-            password,
-          }),
-        }
-      );
-
-      // Redirection vers la page de connexion si la création réussit
-      if (response.status === 200) {
-        navigate("/");
-      } else {
-        // Log des détails de la réponse en cas d'échec
-        setError("Email ou mot de passe incorrect");
-        navigate("/login");
-      }
-    } catch (err) {
-      // Log des erreurs possibles
-      console.error(err);
-    }
-  };
-
+function LoginPage() {
   return (
-    <ThemeProvider theme={defaultTheme}>
-      <Grid container component="main" sx={{ height: "100vh" }}>
-        <CssBaseline />
-        <Grid
-          item
-          xs={false}
-          sm={4}
-          md={7}
-          sx={{
-            backgroundImage:
-              "url(https://img.freepik.com/photos-gratuite/jetee-au-bord-lac-hallstatt-autriche_181624-44201.jpg?w=1380&t=st=1706712919~exp=1706713519~hmac=f269837ae8b59cc38dfcaeaf17ab66e29a4468cfff05b6d0b6cc2abf7117a9c8)",
-            backgroundRepeat: "no-repeat",
-            backgroundColor: (t) =>
-              t.palette.mode === "light"
-                ? t.palette.grey[50]
-                : t.palette.grey[900],
-            backgroundSize: "cover",
-            backgroundPosition: "center",
-          }}
-        />
-        <Grid item xs={12} sm={8} md={5} component={Paper} elevation={0} square>
-          <Box
-            sx={{
-              my: 8,
-              mx: 4,
-              display: "flex",
-              flexDirection: "column",
-              alignItems: "center",
-            }}
-          >
-            <Typography component="h1" variant="h5">
-              Se connecter
-            </Typography>
-            <Box
-              component="form"
-              noValidate
-              onSubmit={handleSubmit}
-              sx={{ mt: 1 }}
-            >
-              <TextField
-                margin="normal"
-                required
-                fullWidth
-                id="email"
-                label="Email Address"
-                name="email"
-                value={email}
-                onChange={handleMailChange}
-                autoFocus
-              />
-              <TextField
-                margin="normal"
-                required
-                fullWidth
-                name="password"
-                label="Password"
-                type="password"
-                id="password"
-                value={password}
-                onChange={handlePasswordChange}
-              />
-              <p style={{ color: "red" }}>{error}</p>
-              <Button
-                type="submit"
-                fullWidth
-                variant="contained"
-                sx={{ mt: 3, mb: 2 }}
-                onClick={handleSubmit}
-              >
-                Se connecter
-              </Button>
-            </Box>
-          </Box>
-        </Grid>
-      </Grid>
-    </ThemeProvider>
+    <div>
+      <Header />
+      <Login />
+      <Navbar />
+    </div>
   );
 }
+
+export default LoginPage;
